+++ packages/core/action_cable_ext/index.d.ts
@@ -6,6 +6,8 @@ export type ExtendedOptions =
       historyTimestamp: number | false
       // Enable PONGs
       pongs: boolean
+      // Disable session recovery completely
+      disableSessionRecovery: boolean
     }>
   | Options